4 step(separate classes into separate files into one class .h and .cpp.)
