+++ Lesson_9/SmartPointer/CarFactory.cpp
@@ -1,0 +1,6 @@
+#include "CarFactory.h"
+
+std::unique_ptr<Car> CarFactory::BuildCar(const std::string& color)
+{
+    return std::unique_ptr<Car>(new Car(color));
+}
