+++ Lesson_9/SmartPointer/CarFactory.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "Car.h"
+
+class CarFactory
+{
+public:
+    std::unique_ptr<Car> BuildCar(const std::string& color);
+};
