+++ Lesson_9/Task.txt
@@ -4,6 +4,9 @@
 4. разнести классы по отдельным файлам на один класс .h и .cpp.
 5. реализовать  SellCar BuyUsedCar. 
 6. реализовать автошколу, которая в отдельном потоке будет производить водителей. 
-7. реализовать класс DriverManager. автошкола произведенного водителя отдает DriverManager-у. DriverManager для каждого нового водителя создает поток, в котором вызывается функция Go
-8. Изменить реализацию функции Driver::Go так, чтобы после ее вызова, Driver периодически вызывал car_->Drive(). Если машины нет - Driver покупает БУ машину, еще через некоторое время покупает новую машину. У кого взять БУ машину может подсказать DriverManager.
+7. реализовать класс DriverManager. автошкола произведенного водителя отдает DriverManager-у. 
+	DriverManager для каждого нового водителя создает поток, в котором вызывается функция Go
+8. Изменить реализацию функции Driver::Go так, чтобы после ее вызова, Driver периодически вызывал car_->Drive(). 
+	Если машины нет - Driver покупает БУ машину, еще через некоторое время покупает новую машину. 
+	У кого взять БУ машину может подсказать DriverManager.
 9. Подумать какие проблемы могут возникнуть после продажи машины и как их решить.
